# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Almacen API",
        "description": "API para administrar items, categorias, proveedores, ubicaciones y movimientos.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/suppliers": {
            "get": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "List suppliers",
                "description": "Display a listing of the resource.",
                "operationId": "43efd859bdde27a9e53c0afbe0891731",
                "responses": {
                    "200": {
                        "description": "Suppliers retrieved successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Create a supplier",
                "description": "Store a newly created resource in storage.",
                "operationId": "c1ee06af9e1f197a4aa49c1dd74c981c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Supplier created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/suppliers/{supplier}": {
            "get": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Show a supplier",
                "description": "Display the specified resource.",
                "operationId": "c1825588a01b42b82a456c6934d8964b",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier retrieved successfully"
                    },
                    "404": {
                        "description": "Supplier not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Update a supplier",
                "description": "Update the specified resource in storage.",
                "operationId": "7ee1b1e4145ec5a9c3615c2202375d30",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supplier updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Delete a supplier",
                "description": "Remove the specified resource from storage.",
                "operationId": "fc078f29b2a734e3fe75d4b5ae80a036",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier deleted successfully"
                    }
                }
            }
        }
    },
...
}